refactor: useActiveTabs를 공용 훅으로 재정의
